ad-process/src/main/java/com/mydomain/adprocess/editing/Editor.java
package com.mydomain.adprocess.editing;

import com.mydomain.adprocess.marketing.Advertisement;
import java.util.Scanner;

public class Editor {

    private static final Object lock = new Object();  // Lock object for synchronization

    @SuppressWarnings("unused")
    private final Scanner scanner;

    public Editor(Scanner scanner) {
        this.scanner = scanner;
    }
    
    // This method simulates the action of an editor reviewing an advertisement's content
    public String reviewContent(Advertisement ad, String userInput) {
        String content = ad.getContent();

        synchronized(lock) {  // Synchronize access to System.in
            Scanner scanner = new Scanner(System.in);
            // Display the content to the editor
            System.out.println("Advertisement Content:");
            System.out.println(content);

            // Prompt the editor for review
            System.out.println("Please review the content and enter 'y' to confirm or 'n' to make changes:");
            String input = scanner.nextLine();

            // Check the editor's input
            if ("y".equalsIgnoreCase(input)) {
                // Content is approved, no changes needed
                System.out.println("Content reviewed and approved.");
            } else if ("n".equalsIgnoreCase(input)) {
                // Content needs changes
                System.out.println("Please enter the updated content:");
                content = scanner.nextLine();  // Get updated content within synchronized block
                System.out.println("Content updated.");
            } else {
                // Invalid input
                System.out.println("Invalid input. Content not changed.");
            }
        }

        // Simulate content editing by appending a review note
        return content + "\n\n[Editor's note: Content reviewed and approved.]";
    }

    // This method simulates adjusting the advertisement's placement based on the magazine's layout
    public String adjustPlacement(Advertisement ad) {
        // Placeholder for actual placement logic
        String placement = ad.getPlacement();
        // For simplicity, we're just confirming the placement without any changes
        return placement;
    }

    // Method to process the entire advertisement
    public Advertisement processAdvertisement(Advertisement ad) {
        String reviewedContent = reviewContent(ad, "userInput"); // replace "userInput" with the actual user input
        String adjustedPlacement = adjustPlacement(ad);

        // We assume that the size and price remain the same, only content and placement
        ad.setContent(reviewedContent);
        ad.setPlacement(adjustedPlacement);

        return ad;
    }
}
