ad-process/src/test/java/com/mydomain/adprocess/editing/EditorTest.java
package com.mydomain.adprocess.editing;

import org.junit.After;
import org.junit.Before;
import org.junit.Test;
import static org.junit.Assert.*;
import java.io.ByteArrayInputStream;
import java.io.InputStream;
import com.mydomain.adprocess.marketing.Advertisement;

public class EditorTest {
    private final InputStream originalSystemIn = System.in;
    private Editor editor;
    private Advertisement ad;

    @Before
    public void setUp() {
        // Provide the input as "y\n" which simulates a user hitting "y" then enter
        String input = "y\n";
        ByteArrayInputStream testIn = new ByteArrayInputStream(input.getBytes());
        System.setIn(testIn);

        // If the Editor class has a no-arg constructor that uses System.in
        editor = new Editor(null);
        ad = new Advertisement("Acme Corp", "Buy Acme products!", "Page 3", 1, 100.0);
    }

    @Test
    public void testReviewContent() {
        String reviewedContent = editor.reviewContent(ad, "y");
        assertTrue("Content should be approved or updated", reviewedContent.contains("[Editor's note: Content reviewed and approved.]"));
    }

    @Test
    public void testAdjustPlacement() {
        String placement = editor.adjustPlacement(ad);
        assertNotNull("Placement should not be null", placement);
    }

    @After
    public void restoreSystemInStream() {
        // Restore original System.in after the test
        System.setIn(originalSystemIn);
    }
}
